Fix README example
