+++ .github/workflows/changelog.yml
@@ -14,6 +14,6 @@ jobs:
       - name: Changelog updated
         uses: Zomzog/changelog-checker@v1.2.0
         with:
-          fileName: CHANGELOG.md
+          fileName: CHANGELOG\.md
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
