+++ .github/workflows/version.yml
@@ -14,6 +14,6 @@ jobs:
       - name: Version updated
         uses: Zomzog/changelog-checker@v1.2.0
         with:
-          fileName: internal/meta/version.go
+          fileName: internal\/meta\/version\.go
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
