+++ test/table_test.go
@@ -4,11 +4,19 @@
 package test
 
 import (
+	"bytes"
 	"context"
+	"fmt"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table/options"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table/resultset"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
+	"log"
 	"math"
 	"os"
+	"path"
 	"sync"
 	"testing"
+	"text/template"
 	"time"
 
 	ydb "github.com/ydb-platform/ydb-go-sdk/v3"
@@ -203,4 +211,650 @@ func TestPoolHealth(t *testing.T) {
 	}()
 
 	wg.Wait()
+}
+
+func driverTrace() trace.Driver {
+	var t trace.Driver
+	trace.Stub(&t, func(name string, args ...interface{}) {
+		log.Printf("[driver] %s: %+v", name, trace.ClearContext(args))
+	})
+	return t
+}
+
+func tableTrace() trace.Table {
+	var t trace.Table
+	trace.Stub(&t, func(name string, args ...interface{}) {
+		log.Printf("[table] %s: %+v", name, trace.ClearContext(args))
+	})
+	return t
+}
+
+func appendConnectOptions(opts ...ydb.Option) []ydb.Option {
+	opts = append(
+		opts,
+		ydb.WithConnectionString(os.Getenv("YDB_CONNECTION_STRING")),
+		ydb.WithTraceDriver(driverTrace()),
+		ydb.WithTraceTable(tableTrace()),
+	)
+	if token, has := os.LookupEnv("YDB_ACCESS_TOKEN_CREDENTIALS"); has {
+		opts = append(opts, ydb.WithAccessTokenCredentials(token))
+	}
+	if v, has := os.LookupEnv("YDB_ANONYMOUS_CREDENTIALS"); has && v == "1" {
+		opts = append(opts, ydb.WithAnonymousCredentials())
+	}
+	return opts
+}
+
+func open(ctx context.Context, opts ...ydb.Option) (ydb.Connection, error) {
+	return ydb.New(ctx, appendConnectOptions(opts...)...)
+}
+
+func seriesData(id uint64, released time.Time, title, info, comment string) types.Value {
+	var commentv types.Value
+	if comment == "" {
+		commentv = types.NullValue(types.TypeUTF8)
+	} else {
+		commentv = types.OptionalValue(types.UTF8Value(comment))
+	}
+	return types.StructValue(
+		types.StructFieldValue("series_id", types.Uint64Value(id)),
+		types.StructFieldValue("release_date", types.DateValueFromTime(released)),
+		types.StructFieldValue("title", types.UTF8Value(title)),
+		types.StructFieldValue("series_info", types.UTF8Value(info)),
+		types.StructFieldValue("comment", commentv),
+	)
+}
+
+func seasonData(seriesID, seasonID uint64, title string, first, last time.Time) types.Value {
+	return types.StructValue(
+		types.StructFieldValue("series_id", types.Uint64Value(seriesID)),
+		types.StructFieldValue("season_id", types.Uint64Value(seasonID)),
+		types.StructFieldValue("title", types.UTF8Value(title)),
+		types.StructFieldValue("first_aired", types.DateValueFromTime(first)),
+		types.StructFieldValue("last_aired", types.DateValueFromTime(last)),
+	)
+}
+
+func episodeData(seriesID, seasonID, episodeID uint64, title string, date time.Time) types.Value {
+	return types.StructValue(
+		types.StructFieldValue("series_id", types.Uint64Value(seriesID)),
+		types.StructFieldValue("season_id", types.Uint64Value(seasonID)),
+		types.StructFieldValue("episode_id", types.Uint64Value(episodeID)),
+		types.StructFieldValue("title", types.UTF8Value(title)),
+		types.StructFieldValue("air_date", types.DateValueFromTime(date)),
+	)
+}
+
+func getSeriesData() types.Value {
+	return types.ListValue(
+		seriesData(
+			1, days("2006-02-03"), "IT Crowd", ""+
+				"The IT Crowd is a British sitcom produced by Channel 4, written by Graham Linehan, produced by "+
+				"Ash Atalla and starring Chris O'Dowd, Richard Ayoade, Katherine Parkinson, and Matt Berry.",
+			"", // NULL comment.
+		),
+		seriesData(
+			2, days("2014-04-06"), "Silicon Valley", ""+
+				"Silicon Valley is an American comedy television series created by Mike Judge, John Altschuler and "+
+				"Dave Krinsky. The series focuses on five young men who founded a startup company in Silicon Valley.",
+			"Some comment here",
+		),
+	)
+}
+
+func getSeasonsData() types.Value {
+	return types.ListValue(
+		seasonData(1, 1, "Season 1", days("2006-02-03"), days("2006-03-03")),
+		seasonData(1, 2, "Season 2", days("2007-08-24"), days("2007-09-28")),
+		seasonData(1, 3, "Season 3", days("2008-11-21"), days("2008-12-26")),
+		seasonData(1, 4, "Season 4", days("2010-06-25"), days("2010-07-30")),
+		seasonData(2, 1, "Season 1", days("2014-04-06"), days("2014-06-01")),
+		seasonData(2, 2, "Season 2", days("2015-04-12"), days("2015-06-14")),
+		seasonData(2, 3, "Season 3", days("2016-04-24"), days("2016-06-26")),
+		seasonData(2, 4, "Season 4", days("2017-04-23"), days("2017-06-25")),
+		seasonData(2, 5, "Season 5", days("2018-03-25"), days("2018-05-13")),
+	)
+}
+
+func getEpisodesData() types.Value {
+	return types.ListValue(
+		episodeData(1, 1, 1, "Yesterday's Jam", days("2006-02-03")),
+		episodeData(1, 1, 2, "Calamity Jen", days("2006-02-03")),
+		episodeData(1, 1, 3, "Fifty-Fifty", days("2006-02-10")),
+		episodeData(1, 1, 4, "The Red Door", days("2006-02-17")),
+		episodeData(1, 1, 5, "The Haunting of Bill Crouse", days("2006-02-24")),
+		episodeData(1, 1, 6, "Aunt Irma Visits", days("2006-03-03")),
+		episodeData(1, 2, 1, "The Work Outing", days("2006-08-24")),
+		episodeData(1, 2, 2, "Return of the Golden Child", days("2007-08-31")),
+		episodeData(1, 2, 3, "Moss and the German", days("2007-09-07")),
+		episodeData(1, 2, 4, "The Dinner Party", days("2007-09-14")),
+		episodeData(1, 2, 5, "Smoke and Mirrors", days("2007-09-21")),
+		episodeData(1, 2, 6, "Men Without Women", days("2007-09-28")),
+		episodeData(1, 3, 1, "From Hell", days("2008-11-21")),
+		episodeData(1, 3, 2, "Are We Not Men?", days("2008-11-28")),
+		episodeData(1, 3, 3, "Tramps Like Us", days("2008-12-05")),
+		episodeData(1, 3, 4, "The Speech", days("2008-12-12")),
+		episodeData(1, 3, 5, "Friendface", days("2008-12-19")),
+		episodeData(1, 3, 6, "Calendar Geeks", days("2008-12-26")),
+		episodeData(1, 4, 1, "Jen The Fredo", days("2010-06-25")),
+		episodeData(1, 4, 2, "The Final Countdown", days("2010-07-02")),
+		episodeData(1, 4, 3, "Something Happened", days("2010-07-09")),
+		episodeData(1, 4, 4, "Italian For Beginners", days("2010-07-16")),
+		episodeData(1, 4, 5, "Bad Boys", days("2010-07-23")),
+		episodeData(1, 4, 6, "Reynholm vs Reynholm", days("2010-07-30")),
+		episodeData(2, 1, 1, "Minimum Viable Product", days("2014-04-06")),
+		episodeData(2, 1, 2, "The Cap Table", days("2014-04-13")),
+		episodeData(2, 1, 3, "Articles of Incorporation", days("2014-04-20")),
+		episodeData(2, 1, 4, "Fiduciary Duties", days("2014-04-27")),
+		episodeData(2, 1, 5, "Signaling Risk", days("2014-05-04")),
+		episodeData(2, 1, 6, "Third Party Insourcing", days("2014-05-11")),
+		episodeData(2, 1, 7, "Proof of Concept", days("2014-05-18")),
+		episodeData(2, 1, 8, "Optimal Tip-to-Tip Efficiency", days("2014-06-01")),
+		episodeData(2, 2, 1, "Sand Hill Shuffle", days("2015-04-12")),
+		episodeData(2, 2, 2, "Runaway Devaluation", days("2015-04-19")),
+		episodeData(2, 2, 3, "Bad Money", days("2015-04-26")),
+		episodeData(2, 2, 4, "The Lady", days("2015-05-03")),
+		episodeData(2, 2, 5, "Server Space", days("2015-05-10")),
+		episodeData(2, 2, 6, "Homicide", days("2015-05-17")),
+		episodeData(2, 2, 7, "Adult Content", days("2015-05-24")),
+		episodeData(2, 2, 8, "White Hat/Black Hat", days("2015-05-31")),
+		episodeData(2, 2, 9, "Binding Arbitration", days("2015-06-07")),
+		episodeData(2, 2, 10, "Two Days of the Condor", days("2015-06-14")),
+		episodeData(2, 3, 1, "Founder Friendly", days("2016-04-24")),
+		episodeData(2, 3, 2, "Two in the Box", days("2016-05-01")),
+		episodeData(2, 3, 3, "Meinertzhagen's Haversack", days("2016-05-08")),
+		episodeData(2, 3, 4, "Maleant Data Systems Solutions", days("2016-05-15")),
+		episodeData(2, 3, 5, "The Empty Chair", days("2016-05-22")),
+		episodeData(2, 3, 6, "Bachmanity Insanity", days("2016-05-29")),
+		episodeData(2, 3, 7, "To Build a Better Beta", days("2016-06-05")),
+		episodeData(2, 3, 8, "Bachman's Earnings Over-Ride", days("2016-06-12")),
+		episodeData(2, 3, 9, "Daily Active Users", days("2016-06-19")),
+		episodeData(2, 3, 10, "The Uptick", days("2016-06-26")),
+		episodeData(2, 4, 1, "Success Failure", days("2017-04-23")),
+		episodeData(2, 4, 2, "Terms of Service", days("2017-04-30")),
+		episodeData(2, 4, 3, "Intellectual Property", days("2017-05-07")),
+		episodeData(2, 4, 4, "Teambuilding Exercise", days("2017-05-14")),
+		episodeData(2, 4, 5, "The Blood Boy", days("2017-05-21")),
+		episodeData(2, 4, 6, "Customer Service", days("2017-05-28")),
+		episodeData(2, 4, 7, "The Patent Troll", days("2017-06-04")),
+		episodeData(2, 4, 8, "The Keenan Vortex", days("2017-06-11")),
+		episodeData(2, 4, 9, "Hooli-Con", days("2017-06-18")),
+		episodeData(2, 4, 10, "Server Error", days("2017-06-25")),
+		episodeData(2, 5, 1, "Grow Fast or Die Slow", days("2018-03-25")),
+		episodeData(2, 5, 2, "Reorientation", days("2018-04-01")),
+		episodeData(2, 5, 3, "Chief Operating Officer", days("2018-04-08")),
+		episodeData(2, 5, 4, "Tech Evangelist", days("2018-04-15")),
+		episodeData(2, 5, 5, "Facial Recognition", days("2018-04-22")),
+		episodeData(2, 5, 6, "Artificial Emotional Intelligence", days("2018-04-29")),
+		episodeData(2, 5, 7, "Initial Coin Offering", days("2018-05-06")),
+		episodeData(2, 5, 8, "Fifty-One Percent", days("2018-05-13")),
+	)
+}
+
+const DateISO8601 = "2006-01-02"
+
+func days(date string) time.Time {
+	t, err := time.Parse(DateISO8601, date)
+	if err != nil {
+		panic(err)
+	}
+	return t
+}
+
+func init() {
+	log.SetFlags(log.LstdFlags | log.Lmicroseconds)
+}
+
+type quet struct {
+}
+
+func (q quet) Write(p []byte) (n int, err error) {
+	return len(p), nil
+}
+
+func Quet() {
+	log.SetOutput(&quet{})
+}
+
+func Prepare(ctx context.Context, db ydb.Connection) error {
+	err := db.Scheme().CleanupDatabase(ctx, db.Name(), "series", "episodes", "seasons")
+	if err != nil {
+		return fmt.Errorf("cleaunup database failed: %w", err)
+	}
+
+	err = db.Scheme().EnsurePathExists(ctx, db.Name())
+	if err != nil {
+		return fmt.Errorf("ensure path exists failed: %w", err)
+	}
+
+	err = describeTableOptions(ctx, db.Table())
+	if err != nil {
+		return fmt.Errorf("describe table options error: %w", err)
+	}
+
+	err = createTables(ctx, db.Table(), db.Name())
+	if err != nil {
+		return fmt.Errorf("create tables error: %w", err)
+	}
+
+	err = describeTable(ctx, db.Table(), path.Join(db.Name(), "series"))
+	if err != nil {
+		return fmt.Errorf("describe table error: %w", err)
+	}
+
+	return nil
+}
+
+func Select(ctx context.Context, db ydb.Connection) error {
+	err := selectSimple(ctx, db.Table(), db.Name())
+	if err != nil {
+		return fmt.Errorf("select simple error: %w", err)
+	}
+
+	err = scanQuerySelect(ctx, db.Table(), db.Name())
+	if err != nil {
+		return fmt.Errorf("scan query error: %w", err)
+	}
+
+	err = readTable(ctx, db.Table(), path.Join(db.Name(), "series"))
+	if err != nil {
+		return fmt.Errorf("read table error: %w", err)
+	}
+
+	return nil
+}
+
+type templateConfig struct {
+	TablePathPrefix string
+}
+
+var fill = template.Must(template.New("fill database").Parse(`
+PRAGMA TablePathPrefix("{{ .TablePathPrefix }}");
+
+DECLARE $seriesData AS List<Struct<
+	series_id: Uint64,
+	title: Utf8,
+	series_info: Utf8,
+	release_date: Date,
+	comment: Optional<Utf8>>>;
+
+DECLARE $seasonsData AS List<Struct<
+	series_id: Uint64,
+	season_id: Uint64,
+	title: Utf8,
+	first_aired: Date,
+	last_aired: Date>>;
+
+DECLARE $episodesData AS List<Struct<
+	series_id: Uint64,
+	season_id: Uint64,
+	episode_id: Uint64,
+	title: Utf8,
+	air_date: Date>>;
+
+REPLACE INTO series
+SELECT
+	series_id,
+	title,
+	series_info,
+	release_date,
+	comment
+FROM AS_TABLE($seriesData);
+
+REPLACE INTO seasons
+SELECT
+	series_id,
+	season_id,
+	title,
+	first_aired,
+	last_aired
+FROM AS_TABLE($seasonsData);
+
+REPLACE INTO episodes
+SELECT
+	series_id,
+	season_id,
+	episode_id,
+	title,
+	air_date
+FROM AS_TABLE($episodesData);
+`))
+
+func readTable(ctx context.Context, c table.Client, path string) error {
+	var (
+		res resultset.Result
+	)
+	err := c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			res, err = s.StreamReadTable(ctx, path,
+				options.ReadOrdered(),
+				options.ReadColumn("series_id"),
+				options.ReadColumn("title"),
+				options.ReadColumn("release_date"),
+			)
+			return err
+		},
+	)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = res.Close()
+	}()
+	var (
+		id    *uint64
+		title *string
+		date  *time.Time
+	)
+	log.Printf("> read_table:\n")
+	for res.NextResultSet(ctx, "series_id", "title", "release_date") {
+		for res.NextRow() {
+			err = res.Scan(&id, &title, &date)
+			if err != nil {
+				return err
+			}
+			log.Printf("  > %d %s %s", *id, *title, date.String())
+		}
+	}
+	if err := res.Err(); err != nil {
+		return err
+	}
+	s := res.Stats()
+	for i := 0; ; i++ {
+		phase, ok := s.NextPhase()
+		if !ok {
+			break
+		}
+		log.Printf(
+			"# phase #%d: took %s",
+			i, phase.Duration(),
+		)
+		for {
+			tbl, ok := phase.NextTableAccess()
+			if !ok {
+				break
+			}
+			log.Printf(
+				"#  accessed %s: read=(%drows, %dbytes)",
+				tbl.Name, tbl.Reads.Rows, tbl.Reads.Bytes,
+			)
+		}
+	}
+	return nil
+}
+
+func describeTableOptions(ctx context.Context, c table.Client) error {
+	var desc options.TableOptionsDescription
+	err := c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			desc, err = s.DescribeTableOptions(ctx)
+			return
+		},
+	)
+	if err != nil {
+		return err
+	}
+	log.Println("> describe_table_options:")
+
+	for i, p := range desc.TableProfilePresets {
+		log.Printf("  > TableProfilePresets: %d/%d: %+v", i+1, len(desc.TableProfilePresets), p)
+	}
+	for i, p := range desc.StoragePolicyPresets {
+		log.Printf("  > StoragePolicyPresets: %d/%d: %+v", i+1, len(desc.StoragePolicyPresets), p)
+	}
+	for i, p := range desc.CompactionPolicyPresets {
+		log.Printf("  > CompactionPolicyPresets: %d/%d: %+v", i+1, len(desc.CompactionPolicyPresets), p)
+	}
+	for i, p := range desc.PartitioningPolicyPresets {
+		log.Printf("  > PartitioningPolicyPresets: %d/%d: %+v", i+1, len(desc.PartitioningPolicyPresets), p)
+	}
+	for i, p := range desc.ExecutionPolicyPresets {
+		log.Printf("  > ExecutionPolicyPresets: %d/%d: %+v", i+1, len(desc.ExecutionPolicyPresets), p)
+	}
+	for i, p := range desc.ReplicationPolicyPresets {
+		log.Printf("  > ReplicationPolicyPresets: %d/%d: %+v", i+1, len(desc.ReplicationPolicyPresets), p)
+	}
+	for i, p := range desc.CachingPolicyPresets {
+		log.Printf("  > CachingPolicyPresets: %d/%d: %+v", i+1, len(desc.CachingPolicyPresets), p)
+	}
+
+	return nil
+}
+
+func selectSimple(ctx context.Context, c table.Client, prefix string) error {
+	var (
+		query = render(
+			template.Must(template.New("").Parse(`
+			PRAGMA TablePathPrefix("{{ .TablePathPrefix }}");
+			DECLARE $seriesID AS Uint64;
+			$format = DateTime::Format("%Y-%m-%d");
+			SELECT
+				series_id,
+				title,
+				release_date
+			FROM
+				series
+			WHERE
+				series_id = $seriesID;
+		`)),
+			templateConfig{
+				TablePathPrefix: prefix,
+			},
+		)
+		res    resultset.Result
+		readTx = table.TxControl(
+			table.BeginTx(
+				table.WithOnlineReadOnly(),
+			),
+			table.CommitTx(),
+		)
+	)
+	err := c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			_, res, err = s.Execute(ctx, readTx, query,
+				table.NewQueryParameters(
+					table.ValueParam("$seriesID", types.Uint64Value(1)),
+				),
+				options.WithQueryCachePolicy(
+					options.WithQueryCachePolicyKeepInCache(),
+				),
+				options.WithCollectStatsModeBasic(),
+			)
+			return err
+		},
+	)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = res.Close()
+	}()
+	var (
+		id    *uint64
+		title *string
+		date  *time.Time
+	)
+
+	log.Printf("> select_simple_transaction:\n")
+	for res.NextResultSet(ctx, "series_id", "title", "release_date") {
+		for res.NextRow() {
+			err = res.Scan(&id, &title, &date)
+			if err != nil {
+				return err
+			}
+			log.Printf(
+				"  > %d %s %s\n",
+				*id, *title, *date,
+			)
+		}
+	}
+	return res.Err()
+}
+
+func scanQuerySelect(ctx context.Context, c table.Client, prefix string) error {
+	var (
+		query = render(
+			template.Must(template.New("").Parse(`
+			PRAGMA TablePathPrefix("{{ .TablePathPrefix }}");
+
+			DECLARE $series AS List<UInt64>;
+
+			SELECT series_id, season_id, title, first_aired
+			FROM seasons
+			WHERE series_id IN $series
+		`)),
+			templateConfig{
+				TablePathPrefix: prefix,
+			},
+		)
+		res resultset.Result
+	)
+	err := c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			res, err = s.StreamExecuteScanQuery(ctx, query,
+				table.NewQueryParameters(
+					table.ValueParam("$series",
+						types.ListValue(
+							types.Uint64Value(1),
+							types.Uint64Value(10),
+						),
+					),
+				),
+			)
+			return err
+		},
+	)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = res.Close()
+	}()
+	var (
+		seriesID uint64
+		seasonID uint64
+		title    string
+		date     time.Time
+	)
+	log.Printf("> scan_query_select:\n")
+	for res.NextResultSet(ctx) {
+		for res.NextRow() {
+			err = res.ScanWithDefaults(&seriesID, &seasonID, &title, &date)
+			if err != nil {
+				return err
+			}
+			log.Printf("  > SeriesId: %d, SeasonId: %d, Title: %s, Air date: %s", seriesID, seasonID, title, date)
+		}
+	}
+	return res.Err()
+}
+
+func Fill(ctx context.Context, db ydb.Connection) error {
+	// Prepare write transaction.
+	writeTx := table.TxControl(
+		table.BeginTx(
+			table.WithSerializableReadWrite(),
+		),
+		table.CommitTx(),
+	)
+	err := db.Table().Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			stmt, err := s.Prepare(ctx, render(fill, templateConfig{
+				TablePathPrefix: db.Name(),
+			}))
+			if err != nil {
+				return
+			}
+			_, _, err = stmt.Execute(ctx, writeTx, table.NewQueryParameters(
+				table.ValueParam("$seriesData", getSeriesData()),
+				table.ValueParam("$seasonsData", getSeasonsData()),
+				table.ValueParam("$episodesData", getEpisodesData()),
+			))
+			return
+		},
+	)
+	return err
+}
+
+func createTables(ctx context.Context, c table.Client, prefix string) error {
+	err := c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			return s.CreateTable(ctx, path.Join(prefix, "series"),
+				options.WithColumn("series_id", types.Optional(types.TypeUint64)),
+				options.WithColumn("title", types.Optional(types.TypeUTF8)),
+				options.WithColumn("series_info", types.Optional(types.TypeUTF8)),
+				options.WithColumn("release_date", types.Optional(types.TypeDate)),
+				options.WithColumn("comment", types.Optional(types.TypeUTF8)),
+				options.WithPrimaryKeyColumn("series_id"),
+			)
+		},
+	)
+	if err != nil {
+		return err
+	}
+
+	err = c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			return s.CreateTable(ctx, path.Join(prefix, "seasons"),
+				options.WithColumn("series_id", types.Optional(types.TypeUint64)),
+				options.WithColumn("season_id", types.Optional(types.TypeUint64)),
+				options.WithColumn("title", types.Optional(types.TypeUTF8)),
+				options.WithColumn("first_aired", types.Optional(types.TypeDate)),
+				options.WithColumn("last_aired", types.Optional(types.TypeDate)),
+				options.WithPrimaryKeyColumn("series_id", "season_id"),
+			)
+		},
+	)
+	if err != nil {
+		return err
+	}
+
+	err = c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			return s.CreateTable(ctx, path.Join(prefix, "episodes"),
+				options.WithColumn("series_id", types.Optional(types.TypeUint64)),
+				options.WithColumn("season_id", types.Optional(types.TypeUint64)),
+				options.WithColumn("episode_id", types.Optional(types.TypeUint64)),
+				options.WithColumn("title", types.Optional(types.TypeUTF8)),
+				options.WithColumn("air_date", types.Optional(types.TypeDate)),
+				options.WithPrimaryKeyColumn("series_id", "season_id", "episode_id"),
+			)
+		},
+	)
+	return err
+}
+
+func describeTable(ctx context.Context, c table.Client, path string) (err error) {
+	err = c.Do(
+		ctx,
+		func(ctx context.Context, s table.Session) (err error) {
+			desc, err := s.DescribeTable(ctx, path)
+			if err != nil {
+				return
+			}
+			log.Printf("> describe table: %s\n", path)
+			for _, c := range desc.Columns {
+				log.Printf("  > column, name: %s, %s\n", c.Type, c.Name)
+			}
+			return
+		},
+	)
+	return err
+}
+
+func render(t *template.Template, data interface{}) string {
+	var buf bytes.Buffer
+	err := t.Execute(&buf, data)
+	if err != nil {
+		panic(err)
+	}
+	return buf.String()
 }
